# public/docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'Cairo Store API'
  description: "كل نقاط المتجر ولوحة التحكم.\r\n\r\nملاحظة عامة على الأخطاء: نقاط JSON في هذا المشروع تُرجع 200 حتى عند\r\nفشل التحقق، والنتيجة تُقرأ من الحقل success في الجسم لا من كود HTTP.\r\nالاستثناءات الوحيدة هي 302 (تحويل لتسجيل الدخول) و403 (صلاحية ناقصة)\r\nو404 (راوت غير موجود). هذا سلوك قائم يعتمد عليه الـfrontend، ووُثّق\r\nهنا كما هو لا كما ينبغي أن يكون."
  version: 1.1.0
servers:
  -
    url: 'http://localhost/STORE/public'
    description: 'Local / Development Server'
paths:
...
  /cart/check-stock:
    post:
      tags:
        - 'Store - Cart'
      summary: 'Check availability and pricing for the variants in the cart'
      description: 'Called from the product pages and the wishlist to refresh their cards with live stock and pricing. Hidden products are not returned.'
      operationId: 11742ec09980c02399157d72f056b111
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              required:
                - variant_ids
              properties:
                variant_ids:
                  description: 'The variant ids to check'
                  type: array
                  items: { type: integer }
              type: object
      responses:
        '200':
          description: "Live stock for the requested variants.\n\nA card rendered on a page that has been left open may be showing a\nstale price and stale stock. This endpoint returns the truth from\nthe database."
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - { properties: { items: { description: 'One entry per requested variant_id. Deleted variants drop out of the result.', type: array, items: { $ref: '#/components/schemas/ProductVariant' } } }, type: object }
  /cart:
    get:
      tags:
        - 'Store - Cart'
      summary: "The current user's cart"
      description: 'Prices and stock are read from the database on every request rather than captured at add time, so any change shows up in the cart before checkout rather than at it.'
      operationId: dd1ea1e457a423ff9cd872a74b1ffa7f
      responses:
        '200':
          description: 'JSON — {success, items, count}'
        '401':
          $ref: '#/components/responses/SessionExpired'
      security:
        -
          userSessionAuth: []
  /cart/add:
    post:
      tags:
        - 'Store - Cart'
      summary: 'Add a quantity to the cart (summed with what is already there)'
      description: 'Adding the same variant twice updates one line rather than creating a second. Stock is not checked here: a cart is an intention, not a reservation, and the reservation happens at /checkout.'
      operationId: 749f5b130bcf852c4a121d2b4618ec31
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - product_id
                - variant_id
                - csrf_token
              properties:
                product_id:
                  type: integer
                variant_id:
                  type: integer
                qty:
                  type: integer
                  default: 1
                  maximum: 100
                  minimum: 1
                csrf_token:
                  type: string
              type: object
      responses:
        '200':
          description: 'JSON — {success, items, count}'
      security:
        -
          userSessionAuth: []
  /cart/update:
    post:
      tags:
        - 'Store - Cart'
      summary: "Set a cart line's quantity (zero removes it)"
      operationId: f151d100ec15be29671e4cafd1eafc6d
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - variant_id
                - qty
                - csrf_token
              properties:
                variant_id:
                  type: integer
                qty:
                  type: integer
                  maximum: 100
                  minimum: 0
                csrf_token:
                  type: string
              type: object
      responses:
        '200':
          description: 'JSON — {success, items, count}'
      security:
        -
          userSessionAuth: []
  /cart/remove:
    post:
      tags:
        - 'Store - Cart'
      summary: 'Remove a line from the cart'
      operationId: eb8d04ff154041f144314563d95d8b00
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - variant_id
                - csrf_token
              properties:
                variant_id:
                  type: integer
                csrf_token:
                  type: string
              type: object
      responses:
        '200':
          description: 'JSON — {success, items, count}'
      security:
        -
          userSessionAuth: []
...
components:
  schemas:
    ApiResponse:
      title: 'غلاف استجابة JSON الموحّد'
      description: "الشكل الذي تُرجعه كل نقطة JSON في المشروع، من Controller::respond().\r\n\r\nكود HTTP يبقى 200 حتى عند الفشل. النتيجة تُقرأ من الحقل success لا\r\nمن كود الحالة. هذا سلوك قائم يعتمد عليه الـfrontend في 34 ملف JS."
      required:
        - success
        - message
      properties:
        success:
          description: 'مصدر الحقيقة الوحيد لنجاح العملية أو فشلها.'
          type: boolean
          example: true
        message:
          description: 'نصّ للعرض على المستخدم. لا تعتمد عليه برمجياً — راجع error_code.'
          type: string
          example: 'تمت العملية بنجاح.'
      type: object
    ApiError:
      title: 'استجابة فشل تحمل رمزاً صريحاً'
      description: "مثل ApiResponse لكن success=false ومعها error_code اختياري.\r\n\r\nالرمز عقد بين الخادم والمتصفح، والنصّ للعرض وحده. الفصل بينهما ليس\r\nتفضيلاً أسلوبياً: كان js/core/csrf.js يكتشف فشل CSRF بمطابقة بداية\r\nنصّ الرسالة، فأي نقطة تصوغ رسالتها بشكل آخر تفقد إعادة المحاولة\r\nالتلقائية بصمت. حدث ذلك ثلاث مرّات قبل أن يُستبدل النصّ برمز."
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: 'Invalid CSRF token, please refresh and try again.'
        error_code:
          description: 'رمز ثابت تقرأه الآلة. يُضاف حين يحتاج العميل التصرّف لا العرض فقط.'
          type: string
          example: csrf_invalid
          enum:
            - csrf_invalid
      type: object
...
    ProductVariant:
      title: 'نسخة لونية من منتج'
      description: 'صفّ من product_variants. لكل منتج variant افتراضي واحد (is_default).'
      properties:
        id:
          type: integer
          example: 7
        product_id:
          type: integer
          example: 42
        color_name:
          type: string
          example: 'Black Titanium'
        color_hex:
          type: string
          example: '#2f2f31'
          nullable: true
        price:
          type: string
          format: decimal
          example: '54999.00'
        discount_percentage:
          type: string
          format: decimal
          example: '0.00'
        price_after_discount:
          type: string
          format: decimal
          nullable: true
        stock_quantity:
          type: integer
          example: 5
        gender_category:
          type: string
          enum:
            - male
            - female
            - both
        image_path:
          type: string
          nullable: true
        is_default:
          type: boolean
          example: true
        sort_order:
          type: integer
          example: 0
      type: object
...
  responses:
...
    SessionExpired:
      description: "لا جلسة صالحة. يُرجعها Middleware::requireAdmin لطلبات AJAX/POST\r\nبكود 401؛ أما طلبات الصفحات الكاملة فتُحوَّل بـ302 إلى صفحة الدخول."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            message: 'Session expired. Please log in again.'
...
  securitySchemes:
    adminSessionAuth:
      type: apiKey
      description: 'PHP session cookie — يُنشأ تلقائياً عند تسجيل الدخول ويُرسل مع كل طلب محمي'
      name: admin_session
      in: cookie
    userSessionAuth:
      type: apiKey
      description: "جلسة المستخدم العادي — منفصلة تماماً عن admin_session بالاسم\r\nوالمحتوى، فلا يمكن الوصول لبيانات الأدمن من جلسة مستخدم ولا العكس\r\n(راجع isUser/isAdmin في auth_helper.php)."
      name: PHPSESSID
      in: cookie
...